Web_App/requirements.txt
absl-py==2.1.0
altair==4.2.2
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
blinker==1.7.0
branca==0.7.0
bs4==0.0.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
entrypoints==0.4
extra-streamlit-components==0.1.60
finance-datareader==0.9.66
flatbuffers==23.5.26
folium==0.15.1
fonttools==4.47.2
gitdb==4.0.11
GitPython==3.1.41
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
mediapipe==0.10.9
mplfinance==0.12.10b0
numpy==1.26.3
opencv-contrib-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.1.4
patsy==0.5.6
pillow==10.2.0
plotly==5.18.0
protobuf==3.20.3
psutil==5.9.7
pyarrow==14.0.2
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
requests==2.31.0
requests-file==1.5.1
rich==13.7.0
rpds-py==0.16.2
scikit-learn==1.3.2
scipy==1.11.4
semver==3.0.2
six==1.16.0
smmap==5.0.1
sounddevice==0.4.6
soupsieve==2.5
soynlp==0.0.493
statsmodels==0.14.1
streamlit==1.19.0
streamlit-authenticator==0.2.3
streamlit-folium==0.18.0
streamlit-lottie==0.0.5
streamlit-option-menu==0.3.6
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.1.0
validators==0.22.0
watchdog==3.0.0
wordcloud==1.9.3
xyzservices==2023.10.1
zipp==3.17.0
